update: we are no longer working with three repositories
As we started development on this reproduction package we were still using three repositories: ours, the quantum-rl one and a third one to generate the csv files, named plot-rl-vqc-data.
At this point, the latter has been integrated in the quantum-rl repository, which has been made public now.
Therefore, we only need two repositories now (ours and quantum-rl).
There have been some relative paths referring to the now unnecessary repository for the generation of csv data, which are being updated by this commit.